feat: Dijkstra routing engine with stop names in segments and ETA fix
- Add Dijkstra in-memory graph routing as primary route calculation engine
- Fix silent NullPointerException in ETA waiting time: R2DBC returned BIGINT
  for COUNT(DISTINCT) column but code requested Integer.class; changed to Long
- Add onErrorResume fallback chain in calculateWaitingTimeFromData so DB errors
  fall through to frequency-based estimation instead of propagating
- Add fromLocationName/toLocationName to RouteSegment; populate stop names
  in DirectRouteOptionBuilder and TransferRouteOptionBuilder
- Update TripOptionDTOConverter to include stop names in location DTOs
- GPS outlier detection improvements and vehicle position updates
- WebFlux config updates, NearbyStops, DirectRouteQuery service changes

+++ src/main/java/biz/ugur/busroutebackend/routing/infrastructure/persistence/repository/R2dbcETARepository.java
@@ -113,7 +113,7 @@ public class R2dbcETARepository implements ETARepository {
                 .map(row -> {
                     Integer baseTime = row.get("base_wait_time", Integer.class);
                     Integer adjustment = row.get("time_adjustment", Integer.class);
-                    Integer vehiclesCount = row.get("vehicles_count", Integer.class);
+                    Long vehiclesCount = row.get("vehicles_count", Long.class);
                     String category = row.get("category", String.class);
 
                     int waitTime = (baseTime != null ? baseTime : 10) + (adjustment != null ? adjustment : 0);
